Better debug logging for Settings
